LPS-56211 rename package
