+++ portal-impl/src/com/liferay/portlet/blogs/action/BaseBlogsImageSelectorUploadHandler.java
@@ -16,7 +16,7 @@ package com.liferay.portlet.blogs.action;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.imageselector.BaseImageSelectorUploadHandler;
+import com.liferay.portal.kernel.image.selector.BaseImageSelectorUploadHandler;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.PortletJSONResponseUtil;
